@@ -1,3 +1,3 @@
-import Strong from './Strong'
+import Strong from "./Strong"
 
 export default Strong
