@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import { Text as RNText } from 'react-native'
 
 import themeManager from '../themeManager'
